+++ RUFAS/routines/EEE/tractor_implement.py
@@ -10,17 +10,19 @@ class TractorImplement:
     def __init__(
         self,
         operation_event: FieldOperationEvent,
+        operation_type: OperationType,
         crop_type: CropType,
         tractor_size: TractorSize,
     ) -> None:
         self.operation_event = operation_event
+        self.operation_type = operation_type
         self.crop_type = crop_type
+        self.tractor_size = tractor_size
         constants = input_manager.get_data("EEE_constants.constants")
         self.constants_by_ID = Utility.convert_list_to_dict_by_key(constants, "ID")
+        self.determine_implement_parameters()
 
-    def determine_implement_parameters(
-        self, crop_type: CropType, tractor_size: TractorSize, operation_type: OperationType
-    ) -> None:
+    def determine_implement_parameters(self) -> None:
         """
         Assign a tractor implement based on the operation, tractor size, and crop type where applicable. Not all
         operations are depended on crop type.
@@ -30,9 +32,9 @@ class TractorImplement:
         dataset = Utility.convert_dict_of_lists_to_list_of_dicts(dataset_raw)
         for data_entry in dataset:
             if (
-                data_entry.get("Crop Type or Tillage Implement").lower() == crop_type.value.lower()
-                and data_entry.get("Tractor Size").lower() == tractor_size.value.lower()
-                and data_entry.get("Operation").lower() == operation_type.value.lower()
+                data_entry.get("Crop Type or Tillage Implement").lower() == self.crop_type.value.lower()
+                and data_entry.get("Tractor Size").lower() == self.tractor_size.value.lower()
+                and data_entry.get("Operation").lower() == self.operation_type.value.lower()
             ):
                 self.implement_name = data_entry["Tractor Implement"]
                 self.A = data_entry["Tractor A (unitless)"]
@@ -46,39 +48,6 @@ class TractorImplement:
                 self.throughput = data_entry["Max Throughput (tons dm/hour)"]
                 self.is_depth_relevant = data_entry["is depth relevant"]
                 break
-
-    def determine_operation_type(self, application_depth: float | None = None) -> List[OperationType]:  # noqa C901
-        """
-        Assigns a specific field operation based on the general name for the operation and the crop type for harvest
-        operations or depth for nutrient application.
-        Implements Helper Function 421 in EEE Functions file.
-        """
-        if self.operation_event == FieldOperationEvent.HARVEST:
-            if self.crop_type in [
-                CropType.ALFALFA_HAY,
-                CropType.ALFALFA_SILAGE,
-                CropType.ALFALFA_BALEAGE,
-                CropType.TALL_FESCUE_HAY,
-                CropType.TALL_FESCUE_SILAGE,
-                CropType.TALL_FESCUE_BALEAGE,
-            ]:
-                return [OperationType.MOWING, OperationType.WINDROWING, OperationType.COLLECTION]
-            else:
-                return [OperationType.COLLECTION]
-        elif self.operation_event == FieldOperationEvent.FERTILIZER_APPLICATION:
-            if application_depth == 0:
-                return [OperationType.FERTILIZER_APPLICATION_SURFACE]
-            elif application_depth > 0:
-                return [OperationType.FERTILIZER_APPLICATION_BELOW_SURFACE]
-        elif self.operation_event == FieldOperationEvent.MANURE_APPLICATION:
-            if application_depth == 0:
-                return [OperationType.LIQUID_MANURE_APPLICATION_SURFACE]
-            elif application_depth > 0:
-                return [OperationType.LIQUID_MANURE_APPLICATION_BELOW_SURFACE]
-        elif self.operation_event == FieldOperationEvent.PLANTING:
-            return [OperationType.PLANTING]
-        elif self.operation_event == FieldOperationEvent.TILLING:
-            return [OperationType.TILLING]
 
     def field_capacity_ha_per_hr(self, crop_yield_ton_per_ha: float | None) -> float:
         """
